requirements.txt
gunpowder == 1.3.4
numpy == 1.26.4
torch == 2.3.1
zarr == 2.18.2
matplotlib == 3.9.2
napari[all] 
funlib-learn-torch @ git+https://github.com/funkelab/funlib.learn.torch.git
PyYAML == 6.0.2
-e .